Refactored suggestion database entry
